Støtter flere perioder per person-vedtak
