+++ .github/workflows/build-and-deploy-preprod.yml
@@ -1,9 +1,6 @@
 name: Build-Deploy-Preprod
-on:
-  push:
-    branches:
-      - '*'
-      - '*/*'
+on: [pull_request]
+
 env:
   IMAGE: docker.pkg.github.com/${{ github.repository }}/familie-ba-sak:${{ github.sha }}
 jobs:
