+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/restDomene/RestVedtakBarn.kt
@@ -11,6 +11,6 @@ data class RestVedtakBarn(
 
 fun VedtakPerson.toRestVedtakBarn() = RestVedtakBarn(
         barn = this.person.personIdent.ident,
-        beløp = this.beløp,
-        stønadFom = this.stønadFom
+        beløp = this.ytelsePerioder.first().beløp, // TODO endre her til å støtte liste
+        stønadFom = this.ytelsePerioder.first().stønadFom
 )
