+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/vedtak/VedtakPerson.kt
@@ -19,24 +19,7 @@ data class VedtakPerson(
         @ManyToOne(optional = false) @JoinColumn(name = "fk_person_id", nullable = false, updatable = false)
         val person: Person,
 
-        @Column(name = "belop", nullable = false)
-        val beløp: Int,
-
-        @Column(name = "stonad_fom", nullable = false)
-        val stønadFom: LocalDate,
-
-        @Column(name = "stonad_tom", nullable = false)
-        val stønadTom: LocalDate,
-
-        @Enumerated(EnumType.STRING)
-        @Column(name = "type", nullable = false)
-        val type: Ytelsetype
+        @OneToMany(fetch = FetchType.EAGER,
+                cascade = [CascadeType.PERSIST, CascadeType.MERGE, CascadeType.REFRESH])
+        val ytelsePerioder: MutableList<YtelsePeriode> = mutableListOf()
 ) : BaseEntitet()
-
-enum class Ytelsetype(val klassifisering: String) {
-    ORDINÆR_BARNETRYGD("BATR"),
-    UTVIDET_BARNETRYGD("BATR"),
-    SMÅBARNSTILLEGG("BATRSMA"),
-    EØS("BATR"),
-    MANUELL_VURDERING("BATR")
-}
