+++ src/main/resources/db/migration/V34__ytelse_periode_migrering.sql
@@ -1,0 +1,2 @@
+INSERT INTO YTELSE_PERIODE (fk_vedtak_person_id, belop, stonad_fom, stonad_tom, type)
+SELECT id, belop, stonad_fom, stonad_tom, type FROM VEDTAK_PERSON;
