+++ src/main/resources/db/migration/V35__delete_vedtak_person.sql
@@ -1,0 +1,4 @@
+ALTER TABLE VEDTAK_PERSON DROP COLUMN stonad_fom,
+DROP COLUMN stonad_tom,
+DROP COLUMN belop,
+DROP COLUMN type;
